Divided RecipeSearchData.js into 3 files
1. Moved functions responsible for modifying JSON objects to new file: PreparingJSONs.js.
2. Moved appKey and appID to new file: AppKeys.js.
3. Added if-else condition in RecipeDataContext.js to make sure that fetching worked correctly in previous step.
4. Increased amount of downloaded recipes from 10 to 100.
