+++ src/data/AppKeys.js
@@ -1,0 +1,2 @@
+export const appID = 'YOUR_APP_ID';    // INSERT YOUR APP ID HERE
+export const appKey = 'YOUR_APP_KEY';  // INSERT YOUR APP KEY HERE
