+++ src/data/RecipeDataContext.js
@@ -8,8 +8,11 @@ export const RecipeDataProvider = ({ children }) => {
     const [currentRecipe, setCurrentRecipe] = useState(null);
 
     useEffect(() => { 
-      downloadRecipes(phrase.length === 0 ? "shrimp" : phrase, 0, 10)
-      .then(recipes => setRecipes(recipes))
+      downloadRecipes(phrase.length === 0 ? "shrimp" : phrase, 0, 100)
+      .then(recipes => {
+          if (recipes === undefined) console.log("Failed to fetch (wrong keys?)"); 
+          else setRecipes(recipes);
+      })
     }, [phrase]);
 
     return (
